Move Cases and Choice to HOL inferences
