+++ Inferences/HOL/Cases.cpp
