+++ Inferences/HOL/Cases.hpp
@@ -17,7 +17,7 @@
 
 #include "Forwards.hpp"
 
-#include "InferenceEngine.hpp"
+#include "Inferences/InferenceEngine.hpp"
 
 namespace Inferences {
