+++ Inferences/HOL/Choice.cpp
